GVN: Invalidate derefs at loop headers
